+++ Tests/Unit/Controller/Fixtures/TestingQueryResult.php
@@ -6,6 +6,7 @@ namespace OliverKlee\Onetimeaccount\Tests\Unit\Controller\Fixtures;
 
 use OliverKlee\FeUserExtraFields\Domain\Model\FrontendUserGroup;
 use TYPO3\CMS\Extbase\Persistence\ObjectStorage;
+use TYPO3\CMS\Extbase\Persistence\QueryInterface;
 use TYPO3\CMS\Extbase\Persistence\QueryResultInterface;
 
 /**
@@ -28,52 +29,52 @@ final class TestingQueryResult implements QueryResultInterface
         $this->objectStorage = $storage;
     }
 
-    public function current()
+    public function current(): FrontendUserGroup
     {
         return $this->objectStorage->current();
     }
 
-    public function next()
+    public function next(): void
     {
         $this->objectStorage->next();
     }
 
-    public function key()
+    public function key(): string
     {
         return $this->objectStorage->key();
     }
 
-    public function valid()
+    public function valid(): bool
     {
         return $this->objectStorage->valid();
     }
 
-    public function rewind()
+    public function rewind(): void
     {
         $this->objectStorage->rewind();
     }
 
-    public function offsetExists($offset)
+    public function offsetExists($offset): bool
     {
         return $this->objectStorage->offsetExists($offset);
     }
 
-    public function offsetGet($offset)
+    public function offsetGet($offset): ?FrontendUserGroup
     {
         return $this->objectStorage->offsetGet($offset);
     }
 
-    public function offsetSet($offset, $value)
+    public function offsetSet($offset, $value): void
     {
         $this->objectStorage->offsetSet($offset, $value);
     }
 
-    public function offsetUnset($offset)
+    public function offsetUnset($offset): void
     {
         $this->objectStorage->offsetUnset($offset);
     }
 
-    public function count()
+    public function count(): int
     {
         return $this->objectStorage->count();
     }
@@ -83,18 +84,21 @@ final class TestingQueryResult implements QueryResultInterface
      *
      * @throws \BadMethodCallException
      */
-    public function getQuery()
+    public function getQuery(): QueryInterface
     {
         throw new \BadMethodCallException('Not implemented.', 1665661687);
     }
 
-    public function getFirst()
+    public function getFirst(): ?FrontendUserGroup
     {
         $this->objectStorage->rewind();
         return $this->objectStorage->current();
     }
 
-    public function toArray()
+    /**
+     * @return array<int, FrontendUserGroup>
+     */
+    public function toArray(): array
     {
         return $this->objectStorage->toArray();
     }